+++ packages/rocketh-export/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rocketh/export
 
+## 0.15.6
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.15.6
+
 ## 0.15.5
 
 ### Patch Changes