+++ packages/rocketh-router/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rocketh/router
 
+## 0.15.7
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.15.6
+  - @rocketh/deploy@0.15.1
+
 ## 0.15.6
 
 ### Patch Changes